requirements.txt
aiofiles==23.1.0
aiohttp==3.8.6
aiosignal==1.3.1
amqp==2.6.1
annotated-types==0.6.0
anyio==4.0.0
apturl==0.5.2
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
bcrypt==3.2.0
billiard==3.5.0.5
blinker==1.4
Brlapi==0.8.3
celery==4.1.0
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.0
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==41.0.4
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.14
diff-match-patch==20230430
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==4.2.6
django-colorfield==0.10.1
django-cors-headers==4.3.0
django-filter==23.3
django-import-export==3.3.3
django-jazzmin==2.6.0
django-oauth-toolkit==2.3.0
django-taggit==3.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-social-oauth2==2.1.3
drf-yasg==1.21.7
duplicity==0.8.21
ecdsa==0.18.0
eskiz-pkg==1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fasteners==0.14.1
frozenlist==1.4.0
future==0.18.2
h11==0.14.0
httpcore==0.18.0
httplib2==0.20.2
httpx==0.25.0
idna==3.4
importlib-metadata==4.6.4
indicator-cpufreq==0.2.2
inflection==0.5.1
jeepney==0.7.1
jwcrypto==1.5.0
keyring==23.5.0
kombu==4.6.11
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
magic-filter==1.0.12
Mako==1.1.3
Markdown==3.5
MarkupPy==1.14
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
netifaces==0.11.0
oauthlib==3.2.2
odfpy==1.4.1
olefile==0.46
openpyxl==3.1.2
packaging==23.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==10.0.1
protobuf==3.12.4
ptyprocess==0.7.0
pyasn1==0.5.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==2.3.0
pydantic_core==2.6.3
PyGObject==3.42.1
PyJWT==2.8.0
pythonPyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pytesseract==0.3.10
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-jose==3.3.0
python-slugify==8.0.1
python-telegram-bot==20.6
python3-openid==3.2.0
pytz==2023.3.post1
pyxdg==0.27
PyYAML==6.0.1
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
systemd-python==234
tablib==3.5.0
text-unidecode==1.3
typing_extensions==4.7.1
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.0.7
usb-creator==0.3.7
vine==1.3.0
wadllib==1.3.6
wrapt==1.15.0
xdg==5
xkit==0.0.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.2
zipp==1.0.0
